+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [GH#223](https://github.com/jolicode/automapper/pull/223) Handle array to Doctrine Collection transformations
 
 ## [9.2.1] - 2025-01-31
 ### Fixed